+++ data/duke.txt
@@ -8,3 +8,4 @@
 [D] | [ ] | return book | 02/12/2019 1800
 [D] | [ ] | return book | 02/12/2019 1800
 [D] | [ ] | return book | 02/12/2019 1800
+[D] | [ ] | return book | 02/12/2019 1800
