+++ src/main/java/duke/DukeException.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Custom exception class for Duke chatbot.
  * This exception throws when issue or error.
